requirements.txt
streamlit
ultralytics
pillow
opencv-python
numpy
